Fixed 2 tasks from TA code review
